+++ src/query/service/tests/it/pipelines/filter/filter_executor.rs
@@ -101,12 +101,8 @@ fn get_filter_data_types() -> Vec<DataType> {
         DataType::Number(NumberDataType::Int64),
         DataType::Number(NumberDataType::Float32),
         DataType::Number(NumberDataType::Float64),
-        DataType::Decimal(DecimalDataType::Decimal128(DecimalSize::new_unchecked(
-            10, 2,
-        ))),
-        DataType::Decimal(DecimalDataType::Decimal128(DecimalSize::new_unchecked(
-            35, 3,
-        ))),
+        DataType::Decimal(DecimalDataType(DecimalSize::new_unchecked(10, 2))),
+        DataType::Decimal(DecimalDataType(DecimalSize::new_unchecked(35, 3))),
         DataType::Array(Box::new(DataType::Number(NumberDataType::UInt32))),
         DataType::Null,
         DataType::Nullable(Box::new(DataType::Number(NumberDataType::UInt32))),
@@ -118,9 +114,7 @@ fn get_filter_data_types() -> Vec<DataType> {
             DataType::Tuple(vec![
                 DataType::Nullable(Box::new(DataType::String)),
                 DataType::Number(NumberDataType::UInt32),
-                DataType::Decimal(DecimalDataType::Decimal128(DecimalSize::new_unchecked(
-                    10, 2,
-                ))),
+                DataType::Decimal(DecimalDataType(DecimalSize::new_unchecked(10, 2))),
             ]),
         ]),
     ]
